remove storage and use ejson storage
